feat: M2-1417 Dashboard: Last Edit time displays in UTC+0 format (Respondents tab)
